Split project endpoints into own files
