+++ src/Language/Mimsa/Server/Project/Evaluate.hs
@@ -1,0 +1,54 @@
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DeriveAnyClass #-}
+{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TypeOperators #-}
+
+module Language.Mimsa.Server.Project.Evaluate
+  ( evaluateExpression,
+    EvaluateAPI,
+  )
+where
+
+import qualified Data.Aeson as JSON
+import Data.Swagger
+import Data.Text (Text)
+import GHC.Generics
+import Language.Mimsa.Printer
+import Language.Mimsa.Server.Handlers
+import Language.Mimsa.Server.Types
+import Language.Mimsa.Types.Project
+import Language.Mimsa.Types.ResolvedExpression
+import Servant
+
+-- /project/evaluate/
+
+type EvaluateAPI =
+  "evaluate" :> ReqBody '[JSON] EvaluateRequest
+    :> Post '[JSON] EvaluateResponse
+
+data EvaluateRequest = EvaluateRequest
+  { erCode :: Text,
+    erProjectHash :: ProjectHash
+  }
+  deriving (Eq, Ord, Show, Generic, JSON.FromJSON, ToSchema)
+
+data EvaluateResponse = EvaluateResponse
+  { erResult :: Text,
+    erExpressionData :: ExpressionData
+  }
+  deriving (Eq, Ord, Show, Generic, JSON.ToJSON, ToSchema)
+
+evaluateExpression ::
+  MimsaEnvironment ->
+  EvaluateRequest ->
+  Handler EvaluateResponse
+evaluateExpression mimsaEnv (EvaluateRequest code hash) = do
+  store' <- readStoreHandler mimsaEnv
+  project <- loadProjectHandler mimsaEnv store' hash
+  (ResolvedExpression mt se expr' scope' swaps) <-
+    evaluateTextHandler project code
+  simpleExpr <-
+    interpretHandler scope' swaps expr'
+  writeStoreHandler mimsaEnv (store project)
+  EvaluateResponse (prettyPrint simpleExpr) <$> expressionDataHandler mimsaEnv se mt
