+++ src/Language/Mimsa/Server/Project/GetExpression.hs
@@ -1,0 +1,53 @@
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DeriveAnyClass #-}
+{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TypeOperators #-}
+
+module Language.Mimsa.Server.Project.GetExpression
+  ( getExpression,
+    GetExpression,
+  )
+where
+
+import qualified Data.Aeson as JSON
+import Data.Swagger
+import GHC.Generics
+import Language.Mimsa.Server.Handlers
+import Language.Mimsa.Server.Types
+import Language.Mimsa.Types.Project
+import Language.Mimsa.Types.ResolvedExpression
+import Language.Mimsa.Types.Store
+import Servant
+
+-- /project/expression/
+
+-- TODO: should this be in Store and have nothing to do with projects?
+-- it could findExpr to get everything we need and then typecheck from there
+type GetExpression =
+  "expression" :> ReqBody '[JSON] GetExpressionRequest
+    :> Post '[JSON] GetExpressionResponse
+
+data GetExpressionRequest = GetExpressionRequest
+  { geProjectHash :: ProjectHash,
+    geExprHash :: ExprHash
+  }
+  deriving (Eq, Ord, Show, Generic, JSON.FromJSON, ToSchema)
+
+newtype GetExpressionResponse = GetExpressionResponse
+  { geExpressionData :: ExpressionData
+  }
+  deriving (Eq, Ord, Show, Generic, JSON.ToJSON, ToSchema)
+
+getExpression ::
+  MimsaEnvironment ->
+  GetExpressionRequest ->
+  Handler GetExpressionResponse
+getExpression mimsaEnv (GetExpressionRequest projectHash exprHash') = do
+  store' <- readStoreHandler mimsaEnv
+  project <- loadProjectHandler mimsaEnv store' projectHash
+  se <- findExprHandler project exprHash'
+  (ResolvedExpression mt _ _ _ _) <-
+    resolveStoreExpressionHandler project se
+  writeStoreHandler mimsaEnv (store project)
+  GetExpressionResponse <$> expressionDataHandler mimsaEnv se mt
